[Weatherinfo] v1.8 improvements
- mode 'OWM': precipitation unit was changed in '%' (no longer in 'mm') like the other services
- some code optimizations

HINTS:
the specification of lon/lat is now also mandatory for MSN
the reduced DICTs are not structured identically. Please note that not all keys exist at all weather services!
function 'getCitylistbyGeocode' is DEPRECATED and will be removed at the end of 2023
function 'getmsnxml' is DEPRECATED and will be removed at the end of 2023