[Weatherinfo] v2.0 update
- bugfix: MSN weathercodes adapted, MSN 'windy-flag' (xxxx1) will be ignored
- new MSN weathercodes added: d603 & n603 "FreezingRainV2"
- some code improvements

HINTS:
the specification of lon/lat is now also mandatory for MSN
the reduced DICTs are not structured identically. Please note that not all keys exist at all weather services!
function 'getCitylistbyGeocode' is DEPRECATED and will be removed at the end of 2023
function 'getmsnxml' is DEPRECATED and will be removed at the end of 2023